+++ rts/Arena.h
@@ -10,13 +10,13 @@
 typedef struct _Arena Arena;
 
 // Start a new arena
-RTS_PRIVATE Arena * newArena   ( void );
+Arena * newArena   ( void );
 
 // Allocate memory in an arena
-RTS_PRIVATE void  * arenaAlloc ( Arena *, size_t );
+void  * arenaAlloc ( Arena *, size_t );
 
 // Free an entire arena
-RTS_PRIVATE void arenaFree  ( Arena * );
+void arenaFree  ( Arena * );
 
 // For internal use only:
 RTS_PRIVATE unsigned long arenaBlocks( void );